+++ lib/slimi.rb
@@ -3,6 +3,7 @@
 require_relative 'slimi/version'
 
 module Slimi
+  autoload :Cli, 'slimi/cli'
   autoload :Engine, 'slimi/engine'
   autoload :Errors, 'slimi/errors'
   autoload :Filters, 'slimi/filters'